reverted splitting of types
